+++ packages/spectacular/src/lib/shared/app-component/spectacular-app.component.spec.ts
@@ -1,4 +1,3 @@
-import { provideLocationMocks } from '@angular/common/testing';
 import { Component } from '@angular/core';
 import { fakeAsync, TestBed, tick } from '@angular/core/testing';
 import { provideRouter, Router } from '@angular/router';
@@ -14,7 +13,6 @@ function setup() {
           component: TestPageComponent,
         },
       ]),
-      provideLocationMocks(),
     ],
   });